requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy<2.0
h5py>=3.8.0
scikit-learn>=1.2.0
scikit-image>=0.20.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Explainability
grad-cam>=1.4.8

# Utilities
psutil>=5.9.0
rich>=13.0.0
pillow>=9.5.0

# Web Server (for inference)
flask>=2.3.0
flask-cors>=4.0.0
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
httpx>=0.25.0

# Hyperparameter Optimization
optuna>=3.3.0
plotly>=5.17.0
